+++ spring-boot-project/spring-boot-actuator-autoconfigure/src/main/java/org/springframework/boot/actuate/autoconfigure/availability/package-info.java
@@ -15,6 +15,7 @@
  */
 
 /**
- * Actuator support for kubernetes-related concerns.
+ * Auto-configuration that extends health endpoints so that they can be used as
+ * availability probes.
  */
-package org.springframework.boot.actuate.kubernetes;
+package org.springframework.boot.actuate.autoconfigure.availability;
